+++ include/mbgl/ios/Mapbox.h
@@ -10,5 +10,6 @@
 #import "MGLPolygon.h"
 #import "MGLPolyline.h"
 #import "MGLShape.h"
+#import "MGLStyle.h"
 #import "MGLTypes.h"
 #import "MGLUserLocation.h"