+++ src/mbgl/util/default_styles.cpp
@@ -1,0 +1,16 @@
+#include <mbgl/util/default_styles.hpp>
+
+namespace mbgl {
+namespace util {
+namespace default_styles {
+
+const DefaultStyle streets   = { "mapbox://styles/mapbox/streets-v8",          "Streets" };
+const DefaultStyle emerald   = { "mapbox://styles/mapbox/emerald-v8",          "Emerald" };
+const DefaultStyle light     = { "mapbox://styles/mapbox/light-v8",            "Light" };
+const DefaultStyle dark      = { "mapbox://styles/mapbox/dark-v8",             "Dark" };
+const DefaultStyle satellite = { "mapbox://styles/mapbox/satellite-v8",        "Satellite" };
+const DefaultStyle hybrid    = { "mapbox://styles/mapbox/satellite-hybrid-v8", "Hybrid" };
+
+}
+} // end namespace util
+} // end namespace mbgl